+++ tests/framework/rest/UrlRuleTest.php
@@ -363,7 +363,7 @@ class UrlRuleTest extends TestCase
     }
 
     /**
-     * @dataProvider testGetCreateUrlStatusProvider
+     * @dataProvider dataProviderGetCreateUrlStatus
      * @param array $config
      * @param array $tests
      */
@@ -400,7 +400,7 @@ class UrlRuleTest extends TestCase
      *   - second element is the expected URL
      *   - third element is the expected result of getCreateUrlStatus() method
      */
-    public function testGetCreateUrlStatusProvider()
+    public function dataProviderGetCreateUrlStatus()
     {
         return [
             'single controller' => [